[DOMA-62][Core] 検索時にselect count(*) を同時に実行する機能を実装する、についてとりあえずコミット。MySQL用の実装がまだできていない。
git-svn-id: https://www.seasar.org/svn/doma/trunk@422 362887e2-e868-44d1-a061-b03550a0b845

+++ doma/src/main/java/org/seasar/doma/internal/jdbc/dialect/HsqldbPagingTransformer.java
@@ -27,7 +27,7 @@ import org.seasar.doma.jdbc.SqlNode;
  */
 public class HsqldbPagingTransformer extends StandardPagingTransformer {
 
-    public HsqldbPagingTransformer(Integer offset, Integer limit) {
+    public HsqldbPagingTransformer(long offset, long limit) {
         super(offset, limit);
     }
